+++ packages/db/tests/search/test-StatsSearch.js
@@ -48,7 +48,8 @@ test.before(async (t) => {
 
   range(20).map((num) => (
     collections.push(fakeCollectionRecordFactory({
-      name: `testCollection___${num}`,
+      name: 'testCollection',
+      version: `${num}`,
       cumulus_id: num,
     }))
   ));
@@ -126,10 +127,7 @@ test.after.always(async (t) => {
 
 test('StatsSearch returns correct response for basic granules query', async (t) => {
   const { knex } = t.context;
-  const queryStringParameters = {
-    type: 'granules',
-  };
-  const AggregateSearch = new StatsSearch({ queryStringParameters }, 'granule');
+  const AggregateSearch = new StatsSearch({}, 'granule');
   const results = await AggregateSearch.aggregate(knex);
   const expectedResponse = [
     { key: 'completed', count: 25 },
@@ -144,7 +142,6 @@ test('StatsSearch returns correct response for basic granules query', async (t) 
 test('StatsSearch filters correctly by date', async (t) => {
   const { knex } = t.context;
   const queryStringParameters = {
-    type: 'granules',
     timestamp__from: `${(new Date(2020, 1, 28)).getTime()}`,
     timestamp__to: `${(new Date(2022, 2, 30)).getTime()}`,
   };
@@ -164,7 +161,6 @@ test('StatsSearch filters correctly by date', async (t) => {
 test('StatsSearch filters executions correctly', async (t) => {
   const { knex } = t.context;
   let queryStringParameters = {
-    type: 'executions',
     field: 'status',
   };
 
@@ -179,7 +175,6 @@ test('StatsSearch filters executions correctly', async (t) => {
   t.deepEqual(results.count, expectedResponse1);
 
   queryStringParameters = {
-    type: 'executions',
     field: 'status',
     timestamp__to: `${(new Date(2023, 11, 30)).getTime()}`,
     timestamp__from: `${(new Date(2021, 1, 28)).getTime()}`,
@@ -196,7 +191,6 @@ test('StatsSearch filters executions correctly', async (t) => {
   t.deepEqual(results2.count, expectedResponse2);
 
   queryStringParameters = {
-    type: 'executions',
     field: 'status',
     timestamp__to: `${(new Date(2023, 11, 30)).getTime()}`,
     timestamp__from: `${(new Date(2021, 1, 28)).getTime()}`,
@@ -214,7 +208,6 @@ test('StatsSearch filters executions correctly', async (t) => {
 test('StatsSearch filters PDRs correctly', async (t) => {
   const { knex } = t.context;
   let queryStringParameters = {
-    type: 'pdrs',
     field: 'status',
   };
 
@@ -229,7 +222,6 @@ test('StatsSearch filters PDRs correctly', async (t) => {
   t.deepEqual(results.count, expectedResponse);
 
   queryStringParameters = {
-    type: 'pdrs',
     field: 'status',
     timestamp__to: `${(new Date(2019, 12, 9)).getTime()}`,
     timestamp__from: `${(new Date(2018, 1, 28)).getTime()}`,
@@ -242,7 +234,6 @@ test('StatsSearch filters PDRs correctly', async (t) => {
   t.deepEqual(results2.count, expectedResponse2);
 
   queryStringParameters = {
-    type: 'pdrs',
     field: 'status',
     timestamp__to: `${(new Date(2019, 12, 9)).getTime()}`,
     timestamp__from: `${(new Date(2018, 1, 28)).getTime()}`,
@@ -259,7 +250,6 @@ test('StatsSearch filters PDRs correctly', async (t) => {
 test('StatsSearch returns correct response when queried by provider', async (t) => {
   const { knex } = t.context;
   const queryStringParameters = {
-    type: 'granules',
     field: 'status',
     provider: 'testProvider2',
   };
@@ -274,7 +264,6 @@ test('StatsSearch returns correct response when queried by provider', async (t) 
 test('StatsSearch returns correct response when queried by collection', async (t) => {
   const { knex } = t.context;
   const queryStringParameters = {
-    type: 'granules',
     field: 'status',
     collectionId: 'testCollection___8',
   };
@@ -289,10 +278,9 @@ test('StatsSearch returns correct response when queried by collection', async (t
 test('StatsSearch returns correct response when queried by collection and provider', async (t) => {
   const { knex } = t.context;
   let queryStringParameters = {
-    type: 'granules',
     field: 'status',
     collectionId: 'testCollection___1',
-    providerId: 'testProvider1',
+    provider: 'testProvider1',
   };
 
   const AggregateSearch = new StatsSearch({ queryStringParameters }, 'granule');
@@ -302,10 +290,9 @@ test('StatsSearch returns correct response when queried by collection and provid
   t.deepEqual(results.count, expectedResponse);
 
   queryStringParameters = {
-    type: 'granules',
     field: 'status',
     collectionId: 'testCollection___1',
-    providerId: 'testProvider1',
+    provider: 'testProvider1',
     timestamp__to: `${(new Date(2019, 12, 9)).getTime()}`,
     timestamp__from: `${(new Date(2018, 1, 28)).getTime()}`,
   };
@@ -316,10 +303,9 @@ test('StatsSearch returns correct response when queried by collection and provid
   t.is(results2.meta.count, 2);
   t.deepEqual(results2.count, expectedResponse2);
   queryStringParameters = {
-    type: 'granules',
     field: 'status',
     collectionId: 'testCollection___1',
-    providerId: 'testProvider1',
+    provider: 'testProvider1',
     timestamp__to: `${(new Date(2019, 12, 9)).getTime()}`,
     timestamp__from: `${(new Date(2018, 1, 28)).getTime()}`,
     status: 'failed',
@@ -335,7 +321,6 @@ test('StatsSearch returns correct response when queried by collection and provid
 test('StatsSearch returns correct response when queried by error', async (t) => {
   const { knex } = t.context;
   let queryStringParameters = {
-    type: 'granules',
     field: 'error.Error.keyword',
   };
   const AggregateSearch = new StatsSearch({ queryStringParameters }, 'granule');
@@ -350,7 +335,6 @@ test('StatsSearch returns correct response when queried by error', async (t) => 
   t.deepEqual(results.count, expectedResponse1);
 
   queryStringParameters = {
-    type: 'granules',
     field: 'error.Error.keyword',
     timestamp__to: `${(new Date(2021, 12, 9)).getTime()}`,
     timestamp__from: `${(new Date(2020, 1, 28)).getTime()}`,
@@ -367,9 +351,8 @@ test('StatsSearch returns correct response when queried by error', async (t) => 
   t.deepEqual(results2.count, expectedResponse2);
 
   queryStringParameters = {
-    type: 'granules',
     collectionId: 'testCollection___1',
-    providerId: 'testProvider1',
+    provider: 'testProvider1',
     field: 'error.Error.keyword',
     timestamp__to: `${(new Date(2019, 12, 9)).getTime()}`,
     timestamp__from: `${(new Date(2018, 1, 28)).getTime()}`,
@@ -384,7 +367,6 @@ test('StatsSearch returns correct response when queried by error', async (t) => 
 test('StatsSearch can query by infix and prefix when type is defined', async (t) => {
   const { knex } = t.context;
   let queryStringParameters = {
-    type: 'granules',
     infix: 'testGra',
   };
   const AggregateSearch = new StatsSearch({ queryStringParameters }, 'granule');
@@ -394,7 +376,6 @@ test('StatsSearch can query by infix and prefix when type is defined', async (t)
   t.deepEqual(results.count, expectedResponse1);
 
   queryStringParameters = {
-    type: 'granules',
     prefix: 'query',
   };
   const AggregateSearch2 = new StatsSearch({ queryStringParameters }, 'granule');
@@ -404,13 +385,13 @@ test('StatsSearch can query by infix and prefix when type is defined', async (t)
   t.deepEqual(results2.count, expectedResponse2);
 
   queryStringParameters = {
-    type: 'collections',
-    infix: 'testCollection___8',
+    infix: 'testCollection',
+    version: '8',
     field: 'name',
   };
   const AggregateSearch3 = new StatsSearch({ queryStringParameters }, 'collection');
   const results3 = await AggregateSearch3.aggregate(knex);
-  const expectedResponse3 = [{ key: 'testCollection___8', count: 1 }];
+  const expectedResponse3 = [{ key: 'testCollection', count: 1 }];
   t.is(results3.meta.count, 1);
   t.deepEqual(results3.count, expectedResponse3);
 });